% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bradleyTerry.R
\name{bradleyTerry}
\alias{bradleyTerry}
\title{Computes the MLE for the BT model using an MM algorithm}
\usage{
bradleyTerry(conf.mat, initial = NA, baseline = NA, stop.dif = 0.001)
}
\arguments{
\item{conf.mat}{an N-by-N matrix. The matrix should be a conflict matrix with element i,j representing the number of times i has beaten j.}

\item{initial}{initial values of dominance indices for the MM algorithm, if not supplied, 1/N (or 0 in the baseline parameterization) 
will be the inital value for all competitors.}

\item{baseline}{index for agent to represent baseline dominance index set to 0.  If NA, the "sum-to-one" parameterization will be used.}

\item{stop.dif}{numeric value for difference in log likelihood value between iterations.  Used as the convergence criterion for the algorithm.}
}
\value{
A list of length 3.
 \item{domInds}{a vector of length N consiting of the MLE values of the dominance indices. Lower values represent lower ranks.}
 \item{probMat}{an N-by-N numeric matrix of win-loss probabilities estimated by the BT model.}
 \item{logLik}{the model fit.}
}
\description{
\code{bradleyTerry} Computes the MLE for the BT model using an MM algorithm
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(sampleEdgelist)
# Computes the MLE for the BT model
bt <- bradleyTerry(confmatrix)
}
\references{
Hunter, R. (2004). MM Algorithm for generalized Bradley Terry Models, The Annals of Statistics 2004, Vol. 32, No. 1, 384–406.

Shev, A., Fujii, K., Hsieh, F., & McCowan, B. (2014). Systemic Testing on Bradley-Terry Model against Nonlinear Ranking Hierarchy. PloS one, 9(12), e115367.
}

