% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bt.test.R
\name{bt.test}
\alias{bt.test}
\title{Systemic test for the assumptions of the Bradley-Terry model}
\usage{
bt.test(conf.mat, baseline = 1, maxLength = 3, reps = 5)
}
\arguments{
\item{conf.mat}{an N-by-N matrix. Either a conflict matrix or a dominance probability matrix (the second element from \code{conductance} output)}

\item{baseline}{an integer between 1 and N inclusive identifying the agent with dominance index equal to zero.}

\item{maxLength}{an integer indicating maximum path length used in conductance}

\item{reps}{an integer indicating number of conflict matrices simulated to estimate the sampling distribution under the BT model.}
}
\value{
A list of length 3.
stat is value of the test statistic.
dist is estimated sampling distribution of the test statistics under the BT model.
p.val is p-value of the test.
}
\description{
\code{bt.test} Systemic test for the assumptions of the Bradley-Terry model, transitivity and monotonic dominance. That is, if A > B and B > C then A > C and Pr(A beats C) > Pr(B beats C).
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(sampleEdgelist)
# test the assumptions of the Bradley-Terry model
condTestoutput <- bt.test(confmatrix)
}

