% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conductance.R
\name{conductance}
\alias{conductance}
\title{compute conductance dominance probabilities}
\usage{
conductance(conf, maxLength, alpha = 6, beta = 1)
}
\arguments{
\item{conf}{N-by-N conflict matrix whose (i,j)th element is the number of times i defeated j}

\item{maxLength}{a positive numeric integer indicating the maximum length of paths to identify}

\item{alpha}{a positive numeric value (more explanation)}

\item{beta}{a positive numeric value (more explanation)}
}
\value{
a list of two elements.
imputed.conf. An N-by-N conflict matrix whose (i,j)th element is the
   'effective' number of wins of i over j.
p.mat. An N-by-N numeric matrix whose (i,j)th element is the estimated
     conductance dominance probability.
}
\description{
\code{conductance} compute conductance dominance probabilities
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# find dominance probability matrix
perm2 <- conductance(confmatrix, 2)
perm2$imputed.conf
perm2$p.hat
}

