% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countPaths.R
\name{countPaths}
\alias{countPaths}
\title{find the number of paths from i to j for each order of dominance path.}
\usage{
countPaths(conf, maxLength = 2)
}
\arguments{
\item{conf}{an N-by-N conflict matrix whose (i,j)th element is the number of times i defeated j}

\item{maxLength}{a positive numeric integer indicating the maximum length of paths to identify}
}
\value{
A list. Elements of the list are number of paths between all pairs of a given length.
}
\description{
\code{countPaths} Identifies the number of paths of length less than or equal to maxLength between all pairs of competitors
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# find number of paths of length 3 or less
npaths <- countPaths(confmatrix, 3)
}

