% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.conflictmat.R
\name{edgelisttomatrix}
\alias{edgelisttomatrix}
\title{transform an edgelist into a matrix}
\usage{
edgelisttomatrix(edgelist, weighted = FALSE, swap.order = FALSE)
}
\arguments{
\item{edgelist}{a 2-column dataframe/matrix of edges. The dominant entity is in the 1st column by default}

\item{swap.order}{If the dominant entity is in the 2nd column, specify as TRUE.}
}
\value{
a named matrix with [i,j]th entry equal to the number of times i dominates j.
}
\description{
transform an edgelist into a matrix
}
\examples{
rawmatrix <- edgelisttomatrix(SampleEdgelist, swap.order = FALSE)

# weighted edgelist
# add example here!
}

