% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findIDpaths.R
\name{findIDpaths}
\alias{findIDpaths}
\title{find all dominance paths of a certain length length for an individual}
\usage{
findIDpaths(conf, ID, len = 2)
}
\arguments{
\item{conf}{an N-by-N conflict matrix whose (i,j)th element is the number of times i defeated j}

\item{ID}{a numeric or character object of length 1. the subject at the beginning of each dominance path.}

\item{len}{a positive integer of length 1. the length of the dominance paths to be identified (len = order + 1)}
}
\value{
return all dominance paths of length (len) beginning at ID
}
\description{
\code{findIDpaths} identifies all unique dominance paths of order (len - 1) beginning at selected subject (given by ID)
}
\examples{
confmatrix <- as.conflictmat(SampleEdgelist)
path38891 <- findIDpaths(confmatrix, ID = 38891, len = 2)
}

