% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findAllRankOrder.R
\name{getAllCosts}
\alias{getAllCosts}
\title{Associate each costs with its corresponding simulated annealing runs}
\usage{
getAllCosts(costs_all, num)
}
\arguments{
\item{costs_all}{costs of all simAnnealing runs. It is the first element of the output from \code{getSimOutput}.}

\item{num}{number of simulated annealing runs}
}
\value{
a data.frame of all costs.
}
\description{
Associate each costs with its corresponding simulated annealing runs
}

