% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findAllRankOrder.R
\name{getAllRankOrder}
\alias{getAllRankOrder}
\title{assign IDs to all best rank orders}
\usage{
getAllRankOrder(ID_index, allRankOrder)
}
\arguments{
\item{ID_index}{it depends on the inputed data from simRankOrder. It takes the colnames of data as ID, and index this ID by its position in the colname.}

\item{allRankOrder}{all rank orders found in all simulated annealing runs. It is the third output from \code{getSimOutput}.}
}
\value{
a data.frame of all costs.
}
\description{
assign IDs to all best rank orders
}

