% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllRankOrder.R
\name{getBestRankOrder}
\alias{getBestRankOrder}
\title{assign IDs to the best rank order}
\usage{
getBestRankOrder(ID_index, bestRankOrder)
}
\arguments{
\item{ID_index}{it depends on the inputed data from simRankOrder. It takes the colnames of data as ID, and index this ID by its position in the colname.}

\item{bestRankOrder}{the best rank order found in all simulated annealing runs. It is the second output from \code{getSimOutput}.}
}
\value{
a data.frame of all costs.
}
\description{
assign IDs to the best rank order
}

