% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllRankOrder.R
\name{getSimOutput}
\alias{getSimOutput}
\title{get useful outputs from simulated annealing processes}
\usage{
getSimOutput(simAnnealList, num)
}
\arguments{
\item{simAnnealList}{the output from simAnnealing process}

\item{num}{number of simulated annealing runs}
}
\value{
a list of three elements

 \item{costs_all}{costs of all simulated annealing runs.}
 \item{bestRankOrder}{best rank order found in all simulated annealing processes}
 \item{allRankOrder}{a dataframe, all best rank orders found in each simulated annealing processes}
}
\description{
get useful outputs from simulated annealing processes
}

