% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/myrecode.R
\name{myrecode}
\alias{myrecode}
\title{recode a vector}
\usage{
myrecode(variable, oldvalue, newvalue)
}
\arguments{
\item{variable}{The vector to be recoded}

\item{oldvalue}{values to be recoded}

\item{newvalue}{values used to replace old values}
}
\value{
A new vector of the same length but replaced the old values using the new values
}
\description{
\code{myrecode} recode values in a vector using new values
}
\details{
old values in new values need to be corresponded to each other.
If you want to replace several values using one value, you need to put it in a list.
}

