% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PercMatrix.R
\name{PercMatrix}
\alias{PercMatrix}
\title{Find the certainty of dominance interactions.}
\usage{
PercMatrix(data, path.length = 2)
}
\arguments{
\item{data}{a data frame. Edgelist. With the first column named "Initiator"; the second column named "Recipient".}

\item{path.length}{an integer between 2 to 4, representing the length of indirect pathways used in finding dominance interactions.}
}
\value{
a dataframe representing dominance certainty matrix.
}
\description{
Find the certainty of dominance interactions.
}
\examples{
PercMatrix(edgelist1)
PercMatrix(edgelist2)
}

