% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.conf.prob.diagnosis.R
\name{plotProbDiagnosis}
\alias{plotProbDiagnosis}
\title{Diagnosis Plot
\code{plotProbDiagnosis} generate heat map for dominance probability matrix}
\usage{
plotProbDiagnosis(prob.mat, cutoff = 0.75, ...)
}
\arguments{
\item{prob.mat}{dominance probability matrix}

\item{cutoff}{a numeric value between 0.5 to 1.
A value that is equal or greater than the cutoff is considered of high certainty.}

\item{...}{Further argument may be supplied and processed by \code{levelplot}.}
}
\description{
Diagnosis Plot
\code{plotProbDiagnosis} generate heat map for dominance probability matrix
}
\seealso{
\code{\link{plotConfmat}}
}

