% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SampleEdgelist.R
\docType{data}
\name{SampleEdgelist}
\alias{SampleEdgelist}
\title{SampleEdgelist.
dominance interactions among 98 monkeys}
\format{A data frame of edgelist with 53940 rows and 2 variables: \code{Initiator}, \code{Recipient}
\describe{
  \item{Initiator}{winner, animal ID}
  \item{Recipient}{loser, animal ID}
  ...
}
McCowan Lab sample data.}
\usage{
SampleEdgelist
}
\description{
SampleEdgelist.
dominance interactions among 98 monkeys
}
\keyword{datasets}

