% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRawMatrix.R
\docType{data}
\name{sampleRawMatrix}
\alias{sampleRawMatrix}
\title{sampleRawMatrix.
dominance interactions between 39 monkeys}
\format{A 39 x 39 matrix representing number of times that a row wins over a column

McCowan Lab sample data.}
\usage{
sampleRawMatrix
}
\description{
sampleRawMatrix.
dominance interactions between 39 monkeys
}
\keyword{datasets}

