% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampleWeightedEdgelist.R
\docType{data}
\name{sampleWeightedEdgelist}
\alias{sampleWeightedEdgelist}
\title{sampleWeightedEdgelist.
dominance interactions among 58 monkeys}
\format{A data frame of edgelist with 50 rows and 3 variables: \code{Initiator}, \code{Recipient}, \code{CountSBT}
\describe{
  \item{Initiator}{winner, animal ID}
  \item{Recipient}{loser, animal ID}
  \item{CountSBT}{Frequency, integer}
  ...
}
McCowan Lab sample data.}
\usage{
sampleWeightedEdgelist
}
\description{
sampleWeightedEdgelist.
dominance interactions among 58 monkeys
}
\keyword{datasets}

