% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.rank.order.R
\name{sim.rank.order}
\alias{sim.rank.order}
\title{Find the rank order}
\usage{
sim.rank.order(data, num = 10, alpha = NULL, kmax = 5)
}
\arguments{
\item{data}{a data frame or a matrix. the win-loss probability matrix
which is the second output from \code{conductance}}

\item{num}{number of SimAnnealing (default is set at 10)}

\item{kmax}{an integer between 2 to 1000}
}
\value{
a dataframe representing simulated rank order.
}
\description{
\code{sim.rank.order} find the rank order for the win-loss relationship
}
\details{
<more information on simAnneal>
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(sampleEdgelist)
# find dominance probability matrix
perm2 <- conductance(confmatrix, maxLength = 2)
s.rank <- sim.rank.order(perm2$p.hat, num = 5, kmax = 5)
head(s.rank)
}
\references{
<add citations here!>
}

