% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conductance.R
\name{valueConverter}
\alias{valueConverter}
\title{dominance probability matrix value converter}
\usage{
valueConverter(matrix)
}
\arguments{
\item{matrix}{the second output from \code{conductance}}
}
\value{
a matrix of win-loss probability ranging from 0.5 - 1.0.
}
\description{
\code{valueConverter} convert values in the dominance probability matrix into 0.5 - 1.0
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# find dominance probability matrix
perm2 <- conductance(confmatrix, 2)
perm2$imputed.conf
perm2$p.hat
convertedValue <- valueConverter(perm2$p.hat)
}

