% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bradleyTerry.R
\name{bradleyTerry}
\alias{bradleyTerry}
\title{Computes the MLE for the BT model using an MM algorithm}
\usage{
bradleyTerry(conf.mat, initial = NA, baseline = NA, stop.dif = 0.001)
}
\arguments{
\item{conf.mat}{an N-by-N matrix. Either a conflict matrix or a dominance probability matrix (the second element from \code{conductance} output)}

\item{initial}{initial values of dominance indices for the MM algorithm, if not supplied, the 0 vector will be the inital value.}

\item{baseline}{index for agent to represent baseline dominance index set to 0.  If NA, the "sum-to-one" parameterization will be used.}

\item{stop.dif}{numeric value for difference in log likelihood value between iterations.  Used as the convergence criterion for the algorithm.}
}
\value{
A list of length 3.
domInds is a vector of length N consiting of the MLE values of the dominance indices.
probMat is an N-by-N numeric matrix of dominance probabilities estimated by the BT model.
logLik is the model fit.
}
\description{
\code{bradleyTerry} Computes the MLE for the BT model using an MM algorithm
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# Computes the MLE for the BT model
bt <- bradleyTerry(confmatrix)
}

