% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bt.test.R
\name{bt.test}
\alias{bt.test}
\title{Systemic test for the assumptions of the Bradley-Terry model}
\usage{
bt.test(conf.mat, baseline = 1, maxLength = 3, reps = 1000)
}
\arguments{
\item{conf.mat}{an N-by-N matrix.
Either a conflict matrix or a win-loss probability matrix
(the second element from \code{conductance} output)}

\item{baseline}{an integer between 1 and N inclusive identifying
the agent with dominance index equal to zero.}

\item{maxLength}{an integer indicating maximum path length used
in \code{conductance}}

\item{reps}{an integer indicating number of conflict matrices
simulated to estimate the sampling distribution under the BT model.}
}
\value{
A list of 3 elements.
 \item{stat}{value of the test statistic}
 \item{dist}{estimated sampling distribution of the test statistics under the BT model.}
 \item{p.val}{p-value of the test}
}
\description{
\code{bt.test} Systemic test for the assumptions of the Bradley-Terry model,
 transitivity and monotonic win-loss relationship.
 That is, if \eqn{A > B} and \eqn{B > C} then \eqn{A > C} and
 \eqn{Pr(A beats C)} > \eqn{Pr(B beats C)}.
}
\details{
The value of the test statistic should be within the estimated
sampling distribution of the test statistics under the BT model.
The p-value of the test indicates the probability of statistics in the estimated sampling distribution
is larger than the test statistic.
It is not appropriate to use Bradley-Terry model if value of the test statistic is higher than
the estimated sampling distribution of the test statistics.
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(sampleEdgelist)
# test the assumptions of the Bradley-Terry model
condTestoutput <- bt.test(confmatrix)
}
\references{
Shev, A., Hsieh, F., Beisner, B., & McCowan, B. (2012). Using Markov chain Monte Carlo (MCMC) to visualize and test the linearity assumption of the Bradley-Terry class of models. Animal behaviour, 84(6), 1523-1531.

 Shev, A., Fujii, K., Hsieh, F., & McCowan, B. (2014). Systemic Testing on Bradley-Terry Model against Nonlinear Ranking Hierarchy. PloS one, 9(12), e115367.
}

