% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conductance.R
\name{conductance}
\alias{conductance}
\title{compute win-loss probabilities}
\usage{
conductance(conf, maxLength, alpha = NULL, beta = 1)
}
\arguments{
\item{conf}{N-by-N conflict matrix whose (i,j)th element is the number of times i defeated j}

\item{maxLength}{a positive numeric integer indicating the maximum length of paths to identify}

\item{alpha}{a positive numeric value (somehow related to the unit of each interaction)}

\item{beta}{a positive numeric value (somehow related to beta distribution)}
}
\value{
a list of two elements.

 \code{imputed.conf} An N-by-N conflict matrix whose (i,j)th element is the
   'effective' number of wins of i over j.

 \code{p.mat} An N-by-N numeric matrix whose (i,j)th element is the estimated
     win-loss probability.
}
\description{
\code{conductance} compute win-loss probabilities for all possible pairs
 based upon the combined information from directed wins/losses and
 indirect win/loss pathways from the network.
}
\details{
<describe this function in more details!>
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(sampleEdgelist)
# find win-loss probability matrix
perm2 <- conductance(confmatrix, 2)
perm2$imputed.conf
perm2$p.hat
}
\references{
<add citations here!>
}

