% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conductance.R
\name{dyadicLongConverter}
\alias{dyadicLongConverter}
\title{dyadic long format converter}
\usage{
dyadicLongConverter(matrix)
}
\arguments{
\item{matrix}{the second output from \code{conductance}}
}
\value{
a dataframe of dyadic level win-loss probability.
}
\description{
\code{dyadicLongConverter} convert dominance probability matrix into long format for each dyad
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# find dominance probability matrix
perm2 <- conductance(confmatrix, 2)
perm2$imputed.conf
perm2$p.hat
dl <- dyadicLongConverter(perm2$p.hat)
}

