% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.conflictmat.R
\name{edgelisttomatrix}
\alias{edgelisttomatrix}
\title{transform an edgelist into a matrix}
\usage{
edgelisttomatrix(edgelist, weighted = FALSE, swap.order = FALSE)
}
\arguments{
\item{edgelist}{a 2-column (or 3-column for weighted edgelist) dataframe/matrix of edges. The winner is in the 1st column by default. For weighted edgelist, the third column should be the weight.}

\item{weighted}{If the edgelist is a 3-column weighted edgelist, use \code{weighted = TRUE}.}

\item{swap.order}{If the winner is in the 2nd column, specify as \code{TRUE}.}
}
\value{
a named matrix with \code{[i,j]}th entry equal to the number of times \code{i} wins over \code{j}.
It is the matrix representation of the edgelist.
}
\description{
transform an edgelist into a matrix
}
\examples{
rawmatrix <- edgelisttomatrix(sampleEdgelist, swap.order = FALSE)

rawmatrix2 <- edgelisttomatrix(sampleWeightedEdgelist, weighted = TRUE, swap.order = FALSE)
}
\seealso{
\code{\link{conductance}}
}

