% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findAllPaths.R
\name{findAllPaths}
\alias{findAllPaths}
\title{Identifies all paths length less than or equal to a certain length between all pairs of competitors}
\usage{
findAllPaths(conf, maxLength = 2)
}
\arguments{
\item{conf}{an N-by-N conflict matrix whose (i,j)th element is the number of times i defeated j}

\item{maxLength}{a positive numeric integer indicating the maximum length of paths to identify}
}
\value{
A list. Elements of the list are all paths of a given length.
}
\description{
\code{findAllPaths} Identifies all paths length less than or equal to maxLength between all pairs of competitors
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# find all paths of legnth 3
allp.3 <- findAllPaths(confmatrix, 3)
}

