% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleEdgelist.R
\docType{data}
\name{sampleEdgelist}
\alias{sampleEdgelist}
\title{sampleEdgelist.
social interactions among 11 monkeys}
\format{A data frame of edgelist with 174 rows and 2 variables: \code{Iname}, \code{Rname}
\describe{
  \item{Iname}{winner, animal ID}
  \item{Rname}{loser, animal ID}
  ...
}
McCowan Lab sample data.}
\usage{
sampleEdgelist
}
\description{
sampleEdgelist.
social interactions among 11 monkeys
}
\keyword{datasets}

