% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleWeightedEdgelist.R
\docType{data}
\name{sampleWeightedEdgelist}
\alias{sampleWeightedEdgelist}
\title{sampleWeightedEdgelist.
dominance interactions among 29 monkeys}
\format{A data frame of edgelist with 181 rows and 3 variables: \code{Initiator1}, \code{Recipient1}, \code{Freq}
\describe{
  \item{Initiator1}{winner, monkey name}
  \item{Recipient1}{loser, monkey name}
  \item{Freq}{Frequency, count of interaction}
  ...
}
McCowan Lab sample data.}
\usage{
sampleWeightedEdgelist
}
\description{
sampleWeightedEdgelist.
dominance interactions among 29 monkeys
}
\keyword{datasets}

