% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.rank.order.R
\name{sim.rank.order}
\alias{sim.rank.order}
\title{Find the rank order}
\usage{
sim.rank.order(data, kmax = 5)
}
\arguments{
\item{data}{a data frame. the output from PercMatrix}

\item{kmax}{an integer between 2 to 1000}
}
\value{
a dataframe representing simulated rank order.
}
\description{
Find the rank order
}
\examples{
rank <- sim.rank.order(PercOutput, 5)
head(rank)
}

