% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.rank.order.R
\name{simRankOrder}
\alias{simRankOrder}
\title{Find the rank order}
\usage{
simRankOrder(data, num = 10, alpha = NULL, kmax = 5)
}
\arguments{
\item{data}{a data frame or a matrix. the win-loss probability matrix
which is the second output from \code{conductance}}

\item{num}{number of SimAnnealing (default is set at 10)}

\item{alpha}{a positive integer that
reflects the influence of an observed win/loss interaction
on an underlying win-loss probability.
It is used in the calculation of the posterior distribution
for the win-loss probability of \code{i} over \code{j}: \eqn{Beta(\alpha c_{i,j} +\beta, c_{i,j}+\beta)}{Beta*(\alpha * c_ij + \beta, c_ij + \beta)}.
In the absence of expertise to accurately estimate alpha,
it is estimated from the data.}

\item{kmax}{an integer between 2 to 1000}
}
\value{
a dataframe representing simulated rank order.
}
\description{
\code{simRankOrder} find the rank order for the win-loss relationship
}
\details{
<more information on simAnneal>
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(sampleEdgelist)
# find dominance probability matrix
perm2 <- conductance(confmatrix, maxLength = 2)
s.rank <- simRankOrder(perm2$p.hat, num = 5, kmax = 5)
head(s.rank)
}
\references{
<add citations here!>
}

