% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitivity.R
\name{transitivity}
\alias{transitivity}
\title{calculate transitivity measurements for a matrix}
\usage{
transitivity(conf)
}
\arguments{
\item{conf}{an N-by-N conflict matrix whose (i,j)th element is the number of times i defeated j}
}
\value{
A list.

 \code{transitive}  The number of transitive triangles.

 \code{intransitive} The number of intransitive triangles.

 \code{transitivity} The order - 1 transitivity.

 \code{alpha} The value of alpha corresponding to this value of transitivity.
}
\description{
\code{transitivity} Identifies all paths length less than or
equal to \code{maxLength} between all pairs of competitors
}
\details{
<more information on transitivity!>
}
\examples{
# convert an edgelist to conflict matrix
confmatrix <- as.conflictmat(SampleEdgelist)
# transitivity calculation
conftrans <- transitivity(confmatrix)
conftrans$transitive
conftrans$intransitive
conftrans$transitivity
conftrans$alpha
}

